Refactor BayesRRmz into BayesRBase and DenseBayesRRmz
This will allow us to add SparseBayesRRG in the future without
duplicating so much code.
