+++ src/parallelgraph.cpp
@@ -1,11 +1,11 @@
 #include "parallelgraph.hpp"
 
 #include "compression.h"
-#include "BayesRRmz.hpp"
+#include "DenseBayesRRmz.hpp"
 
 #include <iostream>
 
-ParallelGraph::ParallelGraph(BayesRRmz *bayes, size_t maxParallel)
+ParallelGraph::ParallelGraph(DenseBayesRRmz *bayes, size_t maxParallel)
     : AnalysisGraph(bayes, maxParallel)
     , m_graph(new graph)
 {
@@ -16,8 +16,8 @@ ParallelGraph::ParallelGraph(BayesRRmz *bayes, size_t maxParallel)
 
         msg.data.reset(new unsigned char[colSize]);
 
-        extractData(reinterpret_cast<unsigned char *>(m_bayes->m_data.ppBedMap) + m_bayes->m_data.ppbedIndex[msg.marker].pos,
-                    static_cast<unsigned int>(m_bayes->m_data.ppbedIndex[msg.marker].size),
+        extractData(reinterpret_cast<unsigned char *>(m_bayes->m_data->ppBedMap) + m_bayes->m_data->ppbedIndex[msg.marker].pos,
+                    static_cast<unsigned int>(m_bayes->m_data->ppbedIndex[msg.marker].size),
                     msg.data.get(),
                     colSize);
